Added FlightListSerializer and FlightDetailSerializer
Created FlightListSerializer for listing flights with route and airplane represented as string related fields, and crew represented as nested objects. Also, added FlightDetailSerializer for detailed flight view, where route and airplane are serialized as nested objects, and crew is represented as a list of objects. Updated FlightViewSet to utilize these serializers based on the action performed.